Rewrite state handling and passing for the sign up screen
