--- src/libs/InitialForm/types.ts
@@ -1,13 +1,0 @@
-type InitialFormProps = {
-  /** Function used to scroll to the top of the page */
-  scrollPageToTop?: () => void;
-};
-
-type InputHandle = {
-  isInputFocused: () => boolean;
-  clearDataAndFocus: (clearLogin?: boolean) => void;
-};
-
-export type {InputHandle};
-
-export default InitialFormProps;
